share auth plugin
