+++ apps/backend/src/handlers/getCodemodDownloadLink.ts
@@ -4,9 +4,9 @@ import type {
   ApiResponse,
   CodemodDownloadLinkResponse,
 } from "@codemod-com/api-types";
+import type { UserDataPopulatedRequest } from "@codemod-com/auth";
 import type { FastifyReply, RouteHandler } from "fastify";
 import { processHandlerError } from "~/types/errors.js";
-import type { UserDataPopulatedRequest } from "../plugins/authPlugin.js";
 import { parseGetCodemodLatestVersionQuery } from "../schemata/schema.js";
 import { codemodService } from "../services/CodemodService.js";
 import { environment } from "../util.js";
