+++ apps/backend/src/handlers/getCodemodsListHandler.ts
@@ -1,7 +1,7 @@
 import { randomBytes } from "node:crypto";
 import type { CodemodListResponse } from "@codemod-com/api-types";
+import type { UserDataPopulatedRequest } from "@codemod-com/auth";
 import type { RouteHandler } from "fastify";
-import type { UserDataPopulatedRequest } from "../plugins/authPlugin.js";
 import {
   parseClientIdentifierSchema,
   parseListCodemodsQuery,
