+++ apps/backend/src/schemata/env.ts
@@ -51,7 +51,7 @@ export const environmentSchema = object({
   SLACK_TOKEN: string(),
   SLACK_CHANNEL: string(),
   BACKEND_API_URL: optional(string()),
-  AUTH_SERVICE_URL: optional(string()),
+  AUTH_SERVICE_URL: string(),
   MODGPT_SERVICE_URL: optional(string()),
   CODEMOD_AI_SERVICE_URL: optional(string()),
   RUN_SERVICE_URL: optional(string()),
