+++ apps/backend/src/server.ts
@@ -1,5 +1,6 @@
 import { randomBytes } from "node:crypto";
 import type { CodemodListResponse } from "@codemod-com/api-types";
+import { getAuthPlugin } from "@codemod-com/auth";
 import { prisma } from "@codemod-com/database";
 import { decryptWithIv, encryptWithIv } from "@codemod-com/utilities";
 import cors, { type FastifyCorsOptions } from "@fastify/cors";
@@ -19,7 +20,6 @@ import {
   getCodemodsHandler,
 } from "./handlers/getCodemodsHandler.js";
 import { getCodemodsListHandler } from "./handlers/getCodemodsListHandler.js";
-import authPlugin from "./plugins/authPlugin.js";
 import {
   type PublishHandlerResponse,
   publishHandler,
@@ -135,8 +135,10 @@ export const initApp = async (toRegister: FastifyPluginCallback[]) => {
     timeWindow: 60 * 1000, // 1 minute
   });
 
-  await fastify.register(fastifyMultipart);
+  const authPlugin = await getAuthPlugin(environment.AUTH_SERVICE_URL);
   await fastify.register(authPlugin);
+
+  await fastify.register(fastifyMultipart);
 
   for (const plugin of toRegister) {
     await fastify.register(plugin);
