+++ apps/modgpt/src/init.ts
@@ -1,10 +1,10 @@
+import { getAuthPlugin } from "@codemod-com/auth";
 import cors from "@fastify/cors";
 import fastifyMultipart from "@fastify/multipart";
 import fastifyRateLimit from "@fastify/rate-limit";
 import Fastify, { type FastifyPluginCallback } from "fastify";
 import { environment } from "./dev-utils/configs";
 import { corsOptions } from "./dev-utils/cors";
-import authPlugin from "./plugins/authPlugin";
 
 export const initApp = async (toRegister: FastifyPluginCallback[]) => {
   const fastify = Fastify({
@@ -39,6 +39,7 @@ export const initApp = async (toRegister: FastifyPluginCallback[]) => {
     handleProcessExit(0);
   });
 
+  const authPlugin = await getAuthPlugin(environment.AUTH_SERVICE_URL);
   await fastify.register(authPlugin);
 
   await fastify.register(cors, corsOptions);
