+++ packages/auth/README.md
@@ -1,0 +1,1 @@
+# Auth shared utils
