+++ packages/auth/package.json
@@ -1,0 +1,37 @@
+{
+  "name": "@codemod-com/auth",
+  "author": "Codemod, Inc.",
+  "version": "1.0.0",
+  "description": "The package that shares common auth plugin logic to be used across backend services.",
+  "type": "module",
+  "main": "./dist/index.js",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "sideEffects": false,
+  "files": [
+    "./dist",
+    "!**/*.test.js",
+    "!**/*.test.d.ts",
+    "README.md",
+    "tsconfig.json"
+  ],
+  "scripts": {
+    "build": "tsc",
+    "test": "vitest run",
+    "coverage": "vitest run --coverage"
+  },
+  "keywords": [],
+  "license": "Apache-2.0",
+  "devDependencies": {
+    "prettier": "^3.2.5",
+    "ts-node": "10.9.1",
+    "vitest": "^1.0.1"
+  },
+  "dependencies": {
+    "@codemod-com/database": "workspace:*",
+    "@codemod-com/utilities": "workspace:*",
+    "axios": "catalog:",
+    "fastify": "catalog:",
+    "fastify-plugin": "catalog:"
+  }
+}
