+++ packages/auth/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./plugin.js";
