+++ packages/auth/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "include": ["./src"],
+  "exclude": ["./dist", "./build", "./node_modules"],
+  "extends": "@codemod-com/tsconfig/lib.json",
+  "compilerOptions": {
+    "baseUrl": ".",
+    "outDir": "./dist"
+  }
+}
